hotfix: change type from float to int
немного напутал бывает... отдавать бэку должно int, а так в cf graphql float(... чуть попозже отрефаторую все

+++ async_cloudflare_stats.py
@@ -341,7 +341,7 @@ class CloudflareAIStats:
                     neurons = group.get('sum', {}).get('totalNeurons')
                     if neurons is not None:
                         try:
-                            total += float(neurons)
+                            total += int(neurons)
                         except (TypeError, ValueError):
                             pass
             return total
